+++ src/fintrist/__init__.py
@@ -1,19 +1,13 @@
 """
 Start Fintrist, connect to the DB, and expose public methods.
 """
-import mongoengine
+from .connect import connect_db, test_db
 from .settings import Config
 from .services import *
 from .models import *
 from fintrist_lib import get_recipe
 
-__all__ = ('Study','BaseStudy', 'get_study', 'Strategy', 'migrate')
+__all__ = ('Study','BaseStudy', 'get_study', 'Strategy', 'migrate',
+    'connect_db', 'mongoclient', 'test_db')
 
-mongoclient = mongoengine.connect(
-    Config.DATABASE_NAME,
-    host=Config.DB_HOST,
-    port=Config.DB_PORT,
-    username=Config.USERNAME,
-    password=Config.PASSWORD,
-    authentication_source='admin',
-)
+mongoclient = connect_db()